[fix] arrumei o lugar do styleSheet
