+++ frontend/src/screen/cadastro/perfil_pessoal/cadastroNomeP.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
-import { View, Text, StyleSheet, TouchableOpacity, TextInput, Image } from "react-native";
+import { View, Text, TouchableOpacity, TextInput, Image } from "react-native";
+import styles from "./stylesNome";
 
 function TelaCadNome() {
 
@@ -41,53 +42,3 @@ function TelaCadNome() {
 
 export default TelaCadNome;
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#FBFBFB',
-        width: '100%',
-        height: '100%',
-        alignItems: 'center',
-        justifyContent: 'center'
-    },
-
-    cima: {
-        flexDirection: 'row',
-    },
-
-    volta: {
-        marginRight: 250
-    },
-
-    button: {
-        height: 68,
-        width: 300,
-        alignItems: 'center',
-        justifyContent: 'center',
-        marginTop: 50,
-        backgroundColor: '#00AD71',
-        borderRadius: 25
-    },
-
-    textgeral: {
-        marginTop: 55,
-        marginBottom: 7,
-        marginLeft: -130,
-        fontSize: 20
-    },
-
-    image: {
-        height: 270,
-        width: 290
-    },
-
-    input: {
-        backgroundColor: '#FFFFFF',
-        borderRadius: 25,
-        height: 68,
-        width: 300,
-        borderWidth: 2,
-        borderStyle: 'solid',
-        borderColor: '#00AD71'
-    }
-});
